requirements.txt
flask==3.0.0
flask-cors==4.0.0
docker==6.1.3
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
aiohttp==3.9.1
requests==2.31.0
pyyaml==6.0.1

